requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
Pillow
djangorestframework
markdown
django-filter